move all cbor stuff in a sub-cbor module
