+++ wallet-crypto/src/cbor/mod.rs
@@ -1,0 +1,5 @@
+mod spec;
+
+// re-export the spec document as if it was defined here...
+pub use self::spec::*;
+pub mod hs;
